+++ src/num/limb/cmp.rs
@@ -1,0 +1,19 @@
+use crate::num::limb::Limb;
+use std::cmp::Ordering;
+impl PartialEq for Limb {
+    fn eq(&self, other: &Self) -> bool {
+        self.0.eq(&other.0)
+    }
+}
+
+impl Ord for Limb {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.0.cmp(&other.0)
+    }
+}
+
+impl PartialOrd for Limb {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
