+++ src/num/mod.rs
@@ -1,2 +1,11 @@
+mod add;
+mod bit_ops;
+mod cmp;
+mod div;
+mod fmt;
+mod from;
 mod limb;
+mod mul;
+mod rem;
+mod sub;
 mod uint;
